requirements.txt
python-telegram-bot
python-telegram-bot[callback-data]
python-telegram-bot[job-queue]
mysql-connector-python
PyMuPDF
pyzbar
Pillow
dotenv
openpyxl
pandas